chore: cargo fmt + clippy cleanup (zero warnings across all targets)
- Apply rustfmt to the whole crate.
- Fix or auto-fix clippy lints; production lib is warning-free.
- Remove dead code flagged by clippy::bad_bit_mask in SCTP DCEP parsing
  ((t & 0x03) == 0x81/0x82 are unreachable; the (t & 0x03) == 0x01/0x02
  checks already cover ordered+unordered cases).
- Replace a tautological 'u16 <= u16::MAX' gap-block test assert with a
  meaningful start<=end range invariant.
- Make IceTransportInner pub(crate) (private-type-leak lint).
- Allow stylistic lints crate-wide where they are intentional design
  choices (too_many_arguments, type_complexity, field_reassign_with_default,
  result_large_err, large_enum_variant); allow test-only pedantic lints in
  test/example crates.
- Guard intentional lock-across-await test cases locally.

cargo clippy --all-targets: 0 warnings. cargo fmt: clean. 427 lib tests pass.

+++ tests/interop_webrtc.rs
@@ -1,3 +1,12 @@
+// Test/example crate: relax pedantic style lints that are noisy in fixtures.
+#![allow(clippy::field_reassign_with_default)]
+#![allow(clippy::redundant_pattern_matching)]
+#![allow(clippy::while_let_loop)]
+#![allow(clippy::manual_checked_ops)]
+#![allow(clippy::needless_range_loop)]
+#![allow(clippy::explicit_counter_loop)]
+#![allow(clippy::cloned_ref_to_slice_refs)]
+#![allow(clippy::zombie_processes)]
 use anyhow::Result;
 use rustrtc::{MediaKind, RtcConfiguration};
 use rustrtc::{PeerConnection, TransceiverDirection};
@@ -226,7 +235,7 @@ async fn interop_vp8_echo() -> Result<()> {
             let mut data = bytes::BytesMut::with_capacity(4);
             data.extend_from_slice(&u32::to_be_bytes(i));
             let frame = rustrtc::media::VideoFrame {
-                rtp_timestamp: i as u32 * 3000,
+                rtp_timestamp: i * 3000,
                 data: data.freeze(),
                 is_last_packet: true,
                 payload_type: None,
@@ -425,7 +434,7 @@ async fn interop_vp8_echo_with_pli() -> Result<()> {
             let mut data = bytes::BytesMut::with_capacity(4);
             data.extend_from_slice(&u32::to_be_bytes(i));
             let frame = rustrtc::media::VideoFrame {
-                rtp_timestamp: i as u32 * 3000,
+                rtp_timestamp: i * 3000,
                 data: data.freeze(),
                 is_last_packet: true,
                 payload_type: None,
